Refactor country/state seeds
