+++ updates/seed_tables.php
@@ -1,0 +1,262 @@
+<?php namespace RainLab\Location\Updates;
+
+use October\Rain\Database\Updates\Seeder;
+use RainLab\Location\Models\Country;
+
+class SeedAllTables extends Seeder
+{
+    public function run()
+    {
+        Country::insert([
+            ['is_enabled' => 1, 'code' => 'AU', 'calling_code' => 61, 'name' => 'Australia'],
+            ['is_enabled' => 1, 'code' => 'CA', 'calling_code' => 1, 'name' => 'Canada'],
+            ['is_enabled' => 1, 'code' => 'GB', 'calling_code' => 44, 'name' => 'United Kingdom'],
+            ['is_enabled' => 1, 'code' => 'US', 'calling_code' => 1, 'name' => 'United States'],
+
+            ['is_enabled' => 0, 'code' => 'AF', 'calling_code' => 93, 'name' => 'Afghanistan'],
+            ['is_enabled' => 0, 'code' => 'AX', 'calling_code' => 358, 'name' => 'Aland Islands '],
+            ['is_enabled' => 0, 'code' => 'AL', 'calling_code' => 355, 'name' => 'Albania'],
+            ['is_enabled' => 0, 'code' => 'DZ', 'calling_code' => 213, 'name' => 'Algeria'],
+            ['is_enabled' => 0, 'code' => 'AS', 'calling_code' => 1684, 'name' => 'American Samoa'],
+            ['is_enabled' => 0, 'code' => 'AD', 'calling_code' => 376, 'name' => 'Andorra'],
+            ['is_enabled' => 0, 'code' => 'AO', 'calling_code' => 244, 'name' => 'Angola'],
+            ['is_enabled' => 0, 'code' => 'AI', 'calling_code' => 1264, 'name' => 'Anguilla'],
+            ['is_enabled' => 0, 'code' => 'AQ', 'calling_code' => 672, 'name' => 'Antarctica'],
+            ['is_enabled' => 0, 'code' => 'AG', 'calling_code' => 1268, 'name' => 'Antigua and Barbuda'],
+            ['is_enabled' => 0, 'code' => 'AR', 'calling_code' => 54, 'name' => 'Argentina'],
+            ['is_enabled' => 0, 'code' => 'AM', 'calling_code' => 374, 'name' => 'Armenia'],
+            ['is_enabled' => 0, 'code' => 'AW', 'calling_code' => 297, 'name' => 'Aruba'],
+            ['is_enabled' => 0, 'code' => 'AT', 'calling_code' => 43, 'name' => 'Austria'],
+            ['is_enabled' => 0, 'code' => 'AZ', 'calling_code' => 994, 'name' => 'Azerbaijan'],
+            ['is_enabled' => 0, 'code' => 'BS', 'calling_code' => 1242, 'name' => 'Bahamas'],
+            ['is_enabled' => 0, 'code' => 'BH', 'calling_code' => 973, 'name' => 'Bahrain'],
+            ['is_enabled' => 0, 'code' => 'BD', 'calling_code' => 880, 'name' => 'Bangladesh'],
+            ['is_enabled' => 0, 'code' => 'BB', 'calling_code' => 1246, 'name' => 'Barbados'],
+            ['is_enabled' => 0, 'code' => 'BY', 'calling_code' => 375, 'name' => 'Belarus'],
+            ['is_enabled' => 0, 'code' => 'BE', 'calling_code' => 32, 'name' => 'Belgium'],
+            ['is_enabled' => 0, 'code' => 'BZ', 'calling_code' => 501, 'name' => 'Belize'],
+            ['is_enabled' => 0, 'code' => 'BJ', 'calling_code' => 229, 'name' => 'Benin'],
+            ['is_enabled' => 0, 'code' => 'BM', 'calling_code' => 1441, 'name' => 'Bermuda'],
+            ['is_enabled' => 0, 'code' => 'BT', 'calling_code' => 975, 'name' => 'Bhutan'],
+            ['is_enabled' => 0, 'code' => 'BO', 'calling_code' => 591, 'name' => 'Bolivia, Plurinational State of'],
+            ['is_enabled' => 0, 'code' => 'BQ', 'calling_code' => 599, 'name' => 'Bonaire, Sint Eustatius and Saba'],
+            ['is_enabled' => 0, 'code' => 'BA', 'calling_code' => 387, 'name' => 'Bosnia and Herzegovina'],
+            ['is_enabled' => 0, 'code' => 'BW', 'calling_code' => 267, 'name' => 'Botswana'],
+            ['is_enabled' => 0, 'code' => 'BV', 'calling_code' => 55, 'name' => 'Bouvet Island'],
+            ['is_enabled' => 0, 'code' => 'BR', 'calling_code' => 55, 'name' => 'Brazil'],
+            ['is_enabled' => 0, 'code' => 'IO', 'calling_code' => 246, 'name' => 'British Indian Ocean Territory'],
+            ['is_enabled' => 0, 'code' => 'BN', 'calling_code' => 673, 'name' => 'Brunei Darussalam'],
+            ['is_enabled' => 0, 'code' => 'BG', 'calling_code' => 359, 'name' => 'Bulgaria'],
+            ['is_enabled' => 0, 'code' => 'BF', 'calling_code' => 226, 'name' => 'Burkina Faso'],
+            ['is_enabled' => 0, 'code' => 'BI', 'calling_code' => 257, 'name' => 'Burundi'],
+            ['is_enabled' => 0, 'code' => 'KH', 'calling_code' => 855, 'name' => 'Cambodia'],
+            ['is_enabled' => 0, 'code' => 'CM', 'calling_code' => 237, 'name' => 'Cameroon'],
+            ['is_enabled' => 0, 'code' => 'CV', 'calling_code' => 238, 'name' => 'Cape Verde'],
+            ['is_enabled' => 0, 'code' => 'KY', 'calling_code' => 1345, 'name' => 'Cayman Islands'],
+            ['is_enabled' => 0, 'code' => 'CF', 'calling_code' => 236, 'name' => 'Central African Republic'],
+            ['is_enabled' => 0, 'code' => 'TD', 'calling_code' => 235, 'name' => 'Chad'],
+            ['is_enabled' => 0, 'code' => 'CL', 'calling_code' => 56, 'name' => 'Chile'],
+            ['is_enabled' => 0, 'code' => 'CN', 'calling_code' => 86, 'name' => 'China'],
+            ['is_enabled' => 0, 'code' => 'CX', 'calling_code' => 61, 'name' => 'Christmas Island'],
+            ['is_enabled' => 0, 'code' => 'CC', 'calling_code' => 61, 'name' => 'Cocos (Keeling) Islands'],
+            ['is_enabled' => 0, 'code' => 'CO', 'calling_code' => 57, 'name' => 'Colombia'],
+            ['is_enabled' => 0, 'code' => 'KM', 'calling_code' => 269, 'name' => 'Comoros'],
+            ['is_enabled' => 0, 'code' => 'CG', 'calling_code' => 242, 'name' => 'Congo'],
+            ['is_enabled' => 0, 'code' => 'CD', 'calling_code' => 243, 'name' => 'Congo, the Democratic Republic of the'],
+            ['is_enabled' => 0, 'code' => 'CK', 'calling_code' => 682, 'name' => 'Cook Islands'],
+            ['is_enabled' => 0, 'code' => 'CR', 'calling_code' => 506, 'name' => 'Costa Rica'],
+            ['is_enabled' => 0, 'code' => 'CI', 'calling_code' => 225, 'name' => 'Cote d\'Ivoire'],
+            ['is_enabled' => 0, 'code' => 'HR', 'calling_code' => 385, 'name' => 'Croatia'],
+            ['is_enabled' => 0, 'code' => 'CU', 'calling_code' => 53, 'name' => 'Cuba'],
+            ['is_enabled' => 0, 'code' => 'CW', 'calling_code' => 599, 'name' => 'Curaçao'],
+            ['is_enabled' => 0, 'code' => 'CY', 'calling_code' => 357, 'name' => 'Cyprus'],
+            ['is_enabled' => 0, 'code' => 'CZ', 'calling_code' => 420, 'name' => 'Czech Republic'],
+            ['is_enabled' => 0, 'code' => 'DK', 'calling_code' => 45, 'name' => 'Denmark'],
+            ['is_enabled' => 0, 'code' => 'DJ', 'calling_code' => 253, 'name' => 'Djibouti'],
+            ['is_enabled' => 0, 'code' => 'DM', 'calling_code' => 1767, 'name' => 'Dominica'],
+            ['is_enabled' => 0, 'code' => 'DO', 'calling_code' => 1809, 'name' => 'Dominican Republic'],
+            ['is_enabled' => 0, 'code' => 'EC', 'calling_code' => 593, 'name' => 'Ecuador'],
+            ['is_enabled' => 0, 'code' => 'EG', 'calling_code' => 20, 'name' => 'Egypt'],
+            ['is_enabled' => 0, 'code' => 'SV', 'calling_code' => 503, 'name' => 'El Salvador'],
+            ['is_enabled' => 0, 'code' => 'GQ', 'calling_code' => 240, 'name' => 'Equatorial Guinea'],
+            ['is_enabled' => 0, 'code' => 'ER', 'calling_code' => 291, 'name' => 'Eritrea'],
+            ['is_enabled' => 0, 'code' => 'EE', 'calling_code' => 372, 'name' => 'Estonia'],
+            ['is_enabled' => 0, 'code' => 'ET', 'calling_code' => 251, 'name' => 'Ethiopia'],
+            ['is_enabled' => 0, 'code' => 'FK', 'calling_code' => 500, 'name' => 'Falkland Islands (Malvinas)'],
+            ['is_enabled' => 0, 'code' => 'FO', 'calling_code' => 298, 'name' => 'Faroe Islands'],
+            ['is_enabled' => 0, 'code' => 'FI', 'calling_code' => 358, 'name' => 'Finland'],
+            ['is_enabled' => 0, 'code' => 'FJ', 'calling_code' => 679, 'name' => 'Fiji'],
+            ['is_enabled' => 1, 'code' => 'FR', 'calling_code' => 33, 'name' => 'France'],
+            ['is_enabled' => 0, 'code' => 'GF', 'calling_code' => 594, 'name' => 'French Guiana'],
+            ['is_enabled' => 0, 'code' => 'PF', 'calling_code' => 689, 'name' => 'French Polynesia'],
+            ['is_enabled' => 0, 'code' => 'TF', 'calling_code' => 262, 'name' => 'French Southern Territories'],
+            ['is_enabled' => 0, 'code' => 'GA', 'calling_code' => 241, 'name' => 'Gabon'],
+            ['is_enabled' => 0, 'code' => 'GM', 'calling_code' => 220, 'name' => 'Gambia'],
+            ['is_enabled' => 0, 'code' => 'GE', 'calling_code' => 995, 'name' => 'Georgia'],
+            ['is_enabled' => 0, 'code' => 'DE', 'calling_code' => 49, 'name' => 'Germany'],
+            ['is_enabled' => 0, 'code' => 'GH', 'calling_code' => 233, 'name' => 'Ghana'],
+            ['is_enabled' => 0, 'code' => 'GI', 'calling_code' => 350, 'name' => 'Gibraltar'],
+            ['is_enabled' => 0, 'code' => 'GR', 'calling_code' => 30, 'name' => 'Greece'],
+            ['is_enabled' => 0, 'code' => 'GL', 'calling_code' => 299, 'name' => 'Greenland'],
+            ['is_enabled' => 0, 'code' => 'GD', 'calling_code' => 1473, 'name' => 'Grenada'],
+            ['is_enabled' => 0, 'code' => 'GP', 'calling_code' => 590, 'name' => 'Guadeloupe'],
+            ['is_enabled' => 0, 'code' => 'GU', 'calling_code' => 1671, 'name' => 'Guam'],
+            ['is_enabled' => 0, 'code' => 'GT', 'calling_code' => 502, 'name' => 'Guatemala'],
+            ['is_enabled' => 0, 'code' => 'GG', 'calling_code' => 44, 'name' => 'Guernsey'],
+            ['is_enabled' => 0, 'code' => 'GN', 'calling_code' => 224, 'name' => 'Guinea'],
+            ['is_enabled' => 0, 'code' => 'GW', 'calling_code' => 245, 'name' => 'Guinea-Bissau'],
+            ['is_enabled' => 0, 'code' => 'GY', 'calling_code' => 592, 'name' => 'Guyana'],
+            ['is_enabled' => 0, 'code' => 'HT', 'calling_code' => 509, 'name' => 'Haiti'],
+            ['is_enabled' => 0, 'code' => 'HM', 'calling_code' => 672, 'name' => 'Heard Island and McDonald Islands'],
+            ['is_enabled' => 0, 'code' => 'VA', 'calling_code' => 39, 'name' => 'Holy See (Vatican City State)'],
+            ['is_enabled' => 0, 'code' => 'HN', 'calling_code' => 504, 'name' => 'Honduras'],
+            ['is_enabled' => 0, 'code' => 'HK', 'calling_code' => 852, 'name' => 'Hong Kong'],
+            ['is_enabled' => 1, 'code' => 'HU', 'calling_code' => 36, 'name' => 'Hungary'],
+            ['is_enabled' => 0, 'code' => 'IS', 'calling_code' => 354, 'name' => 'Iceland'],
+            ['is_enabled' => 1, 'code' => 'IN', 'calling_code' => 91, 'name' => 'India'],
+            ['is_enabled' => 0, 'code' => 'ID', 'calling_code' => 62, 'name' => 'Indonesia'],
+            ['is_enabled' => 0, 'code' => 'IR', 'calling_code' => 98, 'name' => 'Iran, Islamic Republic of'],
+            ['is_enabled' => 0, 'code' => 'IQ', 'calling_code' => 964, 'name' => 'Iraq'],
+            ['is_enabled' => 1, 'code' => 'IE', 'calling_code' => 353, 'name' => 'Ireland'],
+            ['is_enabled' => 0, 'code' => 'IM', 'calling_code' => 44, 'name' => 'Isle of Man'],
+            ['is_enabled' => 0, 'code' => 'IL', 'calling_code' => 972, 'name' => 'Israel'],
+            ['is_enabled' => 0, 'code' => 'IT', 'calling_code' => 39, 'name' => 'Italy'],
+            ['is_enabled' => 0, 'code' => 'JM', 'calling_code' => 1876, 'name' => 'Jamaica'],
+            ['is_enabled' => 0, 'code' => 'JP', 'calling_code' => 81, 'name' => 'Japan'],
+            ['is_enabled' => 0, 'code' => 'JE', 'calling_code' => 44, 'name' => 'Jersey'],
+            ['is_enabled' => 0, 'code' => 'JO', 'calling_code' => 962, 'name' => 'Jordan'],
+            ['is_enabled' => 0, 'code' => 'KZ', 'calling_code' => 7, 'name' => 'Kazakhstan'],
+            ['is_enabled' => 0, 'code' => 'KE', 'calling_code' => 254, 'name' => 'Kenya'],
+            ['is_enabled' => 0, 'code' => 'KI', 'calling_code' => 686, 'name' => 'Kiribati'],
+            ['is_enabled' => 0, 'code' => 'KP', 'calling_code' => 850, 'name' => 'Korea, Democratic People\'s Republic of'],
+            ['is_enabled' => 0, 'code' => 'KR', 'calling_code' => 82, 'name' => 'Korea, Republic of'],
+            ['is_enabled' => 0, 'code' => 'KW', 'calling_code' => 965, 'name' => 'Kuwait'],
+            ['is_enabled' => 0, 'code' => 'KG', 'calling_code' => 996, 'name' => 'Kyrgyzstan'],
+            ['is_enabled' => 0, 'code' => 'LA', 'calling_code' => 856, 'name' => 'Lao People\'s Democratic Republic'],
+            ['is_enabled' => 0, 'code' => 'LV', 'calling_code' => 371, 'name' => 'Latvia'],
+            ['is_enabled' => 0, 'code' => 'LB', 'calling_code' => 961, 'name' => 'Lebanon'],
+            ['is_enabled' => 0, 'code' => 'LS', 'calling_code' => 266, 'name' => 'Lesotho'],
+            ['is_enabled' => 0, 'code' => 'LR', 'calling_code' => 231, 'name' => 'Liberia'],
+            ['is_enabled' => 0, 'code' => 'LY', 'calling_code' => 218, 'name' => 'Libyan Arab Jamahiriya'],
+            ['is_enabled' => 0, 'code' => 'LI', 'calling_code' => 423, 'name' => 'Liechtenstein'],
+            ['is_enabled' => 0, 'code' => 'LT', 'calling_code' => 370, 'name' => 'Lithuania'],
+            ['is_enabled' => 0, 'code' => 'LU', 'calling_code' => 352, 'name' => 'Luxembourg'],
+            ['is_enabled' => 0, 'code' => 'MO', 'calling_code' => 853, 'name' => 'Macao'],
+            ['is_enabled' => 0, 'code' => 'MK', 'calling_code' => 389, 'name' => 'North Macedonia'],
+            ['is_enabled' => 0, 'code' => 'MG', 'calling_code' => 261, 'name' => 'Madagascar'],
+            ['is_enabled' => 0, 'code' => 'MW', 'calling_code' => 265, 'name' => 'Malawi'],
+            ['is_enabled' => 0, 'code' => 'MY', 'calling_code' => 60, 'name' => 'Malaysia'],
+            ['is_enabled' => 0, 'code' => 'MV', 'calling_code' => 960, 'name' => 'Maldives'],
+            ['is_enabled' => 0, 'code' => 'ML', 'calling_code' => 223, 'name' => 'Mali'],
+            ['is_enabled' => 0, 'code' => 'MT', 'calling_code' => 356, 'name' => 'Malta'],
+            ['is_enabled' => 0, 'code' => 'MH', 'calling_code' => 692, 'name' => 'Marshall Islands'],
+            ['is_enabled' => 0, 'code' => 'MQ', 'calling_code' => 596, 'name' => 'Martinique'],
+            ['is_enabled' => 0, 'code' => 'MR', 'calling_code' => 222, 'name' => 'Mauritania'],
+            ['is_enabled' => 0, 'code' => 'MU', 'calling_code' => 230, 'name' => 'Mauritius'],
+            ['is_enabled' => 0, 'code' => 'YT', 'calling_code' => 262, 'name' => 'Mayotte'],
+            ['is_enabled' => 0, 'code' => 'MX', 'calling_code' => 52, 'name' => 'Mexico'],
+            ['is_enabled' => 0, 'code' => 'FM', 'calling_code' => 691, 'name' => 'Micronesia, Federated States of'],
+            ['is_enabled' => 0, 'code' => 'MD', 'calling_code' => 373, 'name' => 'Moldova, Republic of'],
+            ['is_enabled' => 0, 'code' => 'MC', 'calling_code' => 377, 'name' => 'Monaco'],
+            ['is_enabled' => 0, 'code' => 'MN', 'calling_code' => 976, 'name' => 'Mongolia'],
+            ['is_enabled' => 0, 'code' => 'ME', 'calling_code' => 382, 'name' => 'Montenegro'],
+            ['is_enabled' => 0, 'code' => 'MS', 'calling_code' => 1664, 'name' => 'Montserrat'],
+            ['is_enabled' => 0, 'code' => 'MA', 'calling_code' => 212, 'name' => 'Morocco'],
+            ['is_enabled' => 0, 'code' => 'MZ', 'calling_code' => 258, 'name' => 'Mozambique'],
+            ['is_enabled' => 0, 'code' => 'MM', 'calling_code' => 95, 'name' => 'Myanmar'],
+            ['is_enabled' => 0, 'code' => 'NA', 'calling_code' => 264, 'name' => 'Namibia'],
+            ['is_enabled' => 0, 'code' => 'NR', 'calling_code' => 674, 'name' => 'Nauru'],
+            ['is_enabled' => 0, 'code' => 'NP', 'calling_code' => 977, 'name' => 'Nepal'],
+            ['is_enabled' => 1, 'code' => 'NL', 'calling_code' => 31, 'name' => 'Netherlands'],
+            ['is_enabled' => 0, 'code' => 'NC', 'calling_code' => 687, 'name' => 'New Caledonia'],
+            ['is_enabled' => 1, 'code' => 'NZ', 'calling_code' => 64, 'name' => 'New Zealand'],
+            ['is_enabled' => 0, 'code' => 'NI', 'calling_code' => 505, 'name' => 'Nicaragua'],
+            ['is_enabled' => 0, 'code' => 'NE', 'calling_code' => 227, 'name' => 'Niger'],
+            ['is_enabled' => 0, 'code' => 'NG', 'calling_code' => 234, 'name' => 'Nigeria'],
+            ['is_enabled' => 0, 'code' => 'NU', 'calling_code' => 683, 'name' => 'Niue'],
+            ['is_enabled' => 0, 'code' => 'NF', 'calling_code' => 672, 'name' => 'Norfolk Island'],
+            ['is_enabled' => 0, 'code' => 'MP', 'calling_code' => 1670, 'name' => 'Northern Mariana Islands'],
+            ['is_enabled' => 0, 'code' => 'NO', 'calling_code' => 47, 'name' => 'Norway'],
+            ['is_enabled' => 0, 'code' => 'OM', 'calling_code' => 968, 'name' => 'Oman'],
+            ['is_enabled' => 0, 'code' => 'PK', 'calling_code' => 92, 'name' => 'Pakistan'],
+            ['is_enabled' => 0, 'code' => 'PW', 'calling_code' => 680, 'name' => 'Palau'],
+            ['is_enabled' => 0, 'code' => 'PS', 'calling_code' => 970, 'name' => 'Palestine'],
+            ['is_enabled' => 0, 'code' => 'PA', 'calling_code' => 507, 'name' => 'Panama'],
+            ['is_enabled' => 0, 'code' => 'PG', 'calling_code' => 675, 'name' => 'Papua New Guinea'],
+            ['is_enabled' => 0, 'code' => 'PY', 'calling_code' => 595, 'name' => 'Paraguay'],
+            ['is_enabled' => 0, 'code' => 'PE', 'calling_code' => 51, 'name' => 'Peru'],
+            ['is_enabled' => 0, 'code' => 'PH', 'calling_code' => 63, 'name' => 'Philippines'],
+            ['is_enabled' => 0, 'code' => 'PN', 'calling_code' => 870, 'name' => 'Pitcairn'],
+            ['is_enabled' => 0, 'code' => 'PL', 'calling_code' => 48, 'name' => 'Poland'],
+            ['is_enabled' => 0, 'code' => 'PT', 'calling_code' => 351, 'name' => 'Portugal'],
+            ['is_enabled' => 0, 'code' => 'PR', 'calling_code' => 1, 'name' => 'Puerto Rico'],
+            ['is_enabled' => 0, 'code' => 'QA', 'calling_code' => 974, 'name' => 'Qatar'],
+            ['is_enabled' => 0, 'code' => 'RE', 'calling_code' => 262, 'name' => 'Reunion'],
+            ['is_enabled' => 1, 'code' => 'RO', 'calling_code' => 40, 'name' => 'Romania'],
+            ['is_enabled' => 0, 'code' => 'RU', 'calling_code' => 7, 'name' => 'Russian Federation'],
+            ['is_enabled' => 0, 'code' => 'RW', 'calling_code' => 250, 'name' => 'Rwanda'],
+            ['is_enabled' => 0, 'code' => 'BL', 'calling_code' => 590, 'name' => 'Saint Barthélemy'],
+            ['is_enabled' => 0, 'code' => 'SH', 'calling_code' => 290, 'name' => 'Saint Helena'],
+            ['is_enabled' => 0, 'code' => 'KN', 'calling_code' => 1869, 'name' => 'Saint Kitts and Nevis'],
+            ['is_enabled' => 0, 'code' => 'LC', 'calling_code' => 1758, 'name' => 'Saint Lucia'],
+            ['is_enabled' => 0, 'code' => 'MF', 'calling_code' => 1599, 'name' => 'Saint Martin (French part)'],
+            ['is_enabled' => 0, 'code' => 'PM', 'calling_code' => 508, 'name' => 'Saint Pierre and Miquelon'],
+            ['is_enabled' => 0, 'code' => 'VC', 'calling_code' => 1784, 'name' => 'Saint Vincent and the Grenadines'],
+            ['is_enabled' => 0, 'code' => 'WS', 'calling_code' => 685, 'name' => 'Samoa'],
+            ['is_enabled' => 0, 'code' => 'SM', 'calling_code' => 378, 'name' => 'San Marino'],
+            ['is_enabled' => 0, 'code' => 'ST', 'calling_code' => 239, 'name' => 'Sao Tome and Principe'],
+            ['is_enabled' => 0, 'code' => 'SA', 'calling_code' => 966, 'name' => 'Saudi Arabia'],
+            ['is_enabled' => 0, 'code' => 'SN', 'calling_code' => 221, 'name' => 'Senegal'],
+            ['is_enabled' => 0, 'code' => 'RS', 'calling_code' => 381, 'name' => 'Serbia'],
+            ['is_enabled' => 0, 'code' => 'SC', 'calling_code' => 248, 'name' => 'Seychelles'],
+            ['is_enabled' => 0, 'code' => 'SL', 'calling_code' => 232, 'name' => 'Sierra Leone'],
+            ['is_enabled' => 0, 'code' => 'SG', 'calling_code' => 65, 'name' => 'Singapore'],
+            ['is_enabled' => 0, 'code' => 'SX', 'calling_code' => 599, 'name' => 'Sint Maarten (Dutch part)'],
+            ['is_enabled' => 0, 'code' => 'SK', 'calling_code' => 421, 'name' => 'Slovakia'],
+            ['is_enabled' => 0, 'code' => 'SI', 'calling_code' => 386, 'name' => 'Slovenia'],
+            ['is_enabled' => 0, 'code' => 'SB', 'calling_code' => 677, 'name' => 'Solomon Islands'],
+            ['is_enabled' => 0, 'code' => 'SO', 'calling_code' => 252, 'name' => 'Somalia'],
+            ['is_enabled' => 0, 'code' => 'ZA', 'calling_code' => 27, 'name' => 'South Africa'],
+            ['is_enabled' => 0, 'code' => 'GS', 'calling_code' => 500, 'name' => 'South Georgia and the South Sandwich Islands'],
+            ['is_enabled' => 1, 'code' => 'ES', 'calling_code' => 34, 'name' => 'Spain'],
+            ['is_enabled' => 0, 'code' => 'LK', 'calling_code' => 94, 'name' => 'Sri Lanka'],
+            ['is_enabled' => 0, 'code' => 'SD', 'calling_code' => 249, 'name' => 'Sudan'],
+            ['is_enabled' => 0, 'code' => 'SR', 'calling_code' => 597, 'name' => 'Suriname'],
+            ['is_enabled' => 0, 'code' => 'SJ', 'calling_code' => 47, 'name' => 'Svalbard and Jan Mayen'],
+            ['is_enabled' => 0, 'code' => 'SZ', 'calling_code' => 268, 'name' => 'Swaziland'],
+            ['is_enabled' => 0, 'code' => 'SE', 'calling_code' => 46, 'name' => 'Sweden'],
+            ['is_enabled' => 0, 'code' => 'CH', 'calling_code' => 41, 'name' => 'Switzerland'],
+            ['is_enabled' => 0, 'code' => 'SY', 'calling_code' => 963, 'name' => 'Syrian Arab Republic'],
+            ['is_enabled' => 0, 'code' => 'TW', 'calling_code' => 886, 'name' => 'Taiwan, Province of China'],
+            ['is_enabled' => 0, 'code' => 'TJ', 'calling_code' => 992, 'name' => 'Tajikistan'],
+            ['is_enabled' => 0, 'code' => 'TZ', 'calling_code' => 255, 'name' => 'Tanzania, United Republic of'],
+            ['is_enabled' => 0, 'code' => 'TH', 'calling_code' => 66, 'name' => 'Thailand'],
+            ['is_enabled' => 0, 'code' => 'TL', 'calling_code' => 670, 'name' => 'Timor-Leste'],
+            ['is_enabled' => 0, 'code' => 'TG', 'calling_code' => 228, 'name' => 'Togo'],
+            ['is_enabled' => 0, 'code' => 'TK', 'calling_code' => 690, 'name' => 'Tokelau'],
+            ['is_enabled' => 0, 'code' => 'TO', 'calling_code' => 676, 'name' => 'Tonga'],
+            ['is_enabled' => 0, 'code' => 'TT', 'calling_code' => 1868, 'name' => 'Trinidad and Tobago'],
+            ['is_enabled' => 0, 'code' => 'TN', 'calling_code' => 216, 'name' => 'Tunisia'],
+            ['is_enabled' => 0, 'code' => 'TR', 'calling_code' => 90, 'name' => 'Turkey'],
+            ['is_enabled' => 0, 'code' => 'TM', 'calling_code' => 993, 'name' => 'Turkmenistan'],
+            ['is_enabled' => 0, 'code' => 'TC', 'calling_code' => 1649, 'name' => 'Turks and Caicos Islands'],
+            ['is_enabled' => 0, 'code' => 'TV', 'calling_code' => 688, 'name' => 'Tuvalu'],
+            ['is_enabled' => 0, 'code' => 'UG', 'calling_code' => 256, 'name' => 'Uganda'],
+            ['is_enabled' => 0, 'code' => 'UA', 'calling_code' => 380, 'name' => 'Ukraine'],
+            ['is_enabled' => 0, 'code' => 'AE', 'calling_code' => 971, 'name' => 'United Arab Emirates'],
+            ['is_enabled' => 0, 'code' => 'UM', 'calling_code' => 246, 'name' => 'United States Minor Outlying Islands'],
+            ['is_enabled' => 0, 'code' => 'UY', 'calling_code' => 598, 'name' => 'Uruguay'],
+            ['is_enabled' => 0, 'code' => 'UZ', 'calling_code' => 998, 'name' => 'Uzbekistan'],
+            ['is_enabled' => 0, 'code' => 'VU', 'calling_code' => 678, 'name' => 'Vanuatu'],
+            ['is_enabled' => 0, 'code' => 'VE', 'calling_code' => 58, 'name' => 'Venezuela, Bolivarian Republic of'],
+            ['is_enabled' => 0, 'code' => 'VN', 'calling_code' => 84, 'name' => 'Viet Nam'],
+            ['is_enabled' => 0, 'code' => 'VG', 'calling_code' => 1284, 'name' => 'Virgin Islands, British'],
+            ['is_enabled' => 0, 'code' => 'VI', 'calling_code' => 1340, 'name' => 'Virgin Islands, U.S.'],
+            ['is_enabled' => 0, 'code' => 'WF', 'calling_code' => 681, 'name' => 'Wallis and Futuna'],
+            ['is_enabled' => 0, 'code' => 'EH', 'calling_code' => 212, 'name' => 'Western Sahara'],
+            ['is_enabled' => 0, 'code' => 'YE', 'calling_code' => 967, 'name' => 'Yemen'],
+            ['is_enabled' => 0, 'code' => 'ZM', 'calling_code' => 260, 'name' => 'Zambia'],
+            ['is_enabled' => 0, 'code' => 'ZW', 'calling_code' => 263, 'name' => 'Zimbabwe'],
+        ]);
+    }
+}
