+++ updates/seed_us_states.php
@@ -1,0 +1,77 @@
+<?php namespace RainLab\Location\Updates;
+
+use October\Rain\Database\Updates\Seeder;
+use RainLab\Location\Models\Country;
+
+class SeedUsStates extends Seeder
+{
+    public function run()
+    {
+        $country = Country::whereCode('US')->first();
+        if ($country->states()->count() > 0) {
+            return;
+        }
+
+        $country->states()->createMany([
+            ['code' => 'AL', 'name' => 'Alabama'],
+            ['code' => 'AK', 'name' => 'Alaska'],
+            ['code' => 'AS', 'name' => 'American Samoa'],
+            ['code' => 'AZ', 'name' => 'Arizona'],
+            ['code' => 'AR', 'name' => 'Arkansas'],
+            ['code' => 'CA', 'name' => 'California'],
+            ['code' => 'CO', 'name' => 'Colorado'],
+            ['code' => 'CT', 'name' => 'Connecticut'],
+            ['code' => 'DE', 'name' => 'Delaware'],
+            ['code' => 'DC', 'name' => 'Dist. of Columbia'],
+            ['code' => 'FL', 'name' => 'Florida'],
+            ['code' => 'GA', 'name' => 'Georgia'],
+            ['code' => 'GU', 'name' => 'Guam'],
+            ['code' => 'HI', 'name' => 'Hawaii'],
+            ['code' => 'ID', 'name' => 'Idaho'],
+            ['code' => 'IL', 'name' => 'Illinois'],
+            ['code' => 'IN', 'name' => 'Indiana'],
+            ['code' => 'IA', 'name' => 'Iowa'],
+            ['code' => 'KS', 'name' => 'Kansas'],
+            ['code' => 'KY', 'name' => 'Kentucky'],
+            ['code' => 'LA', 'name' => 'Louisiana'],
+            ['code' => 'ME', 'name' => 'Maine'],
+            ['code' => 'MD', 'name' => 'Maryland'],
+            ['code' => 'MH', 'name' => 'Marshall Islands'],
+            ['code' => 'MA', 'name' => 'Massachusetts'],
+            ['code' => 'MI', 'name' => 'Michigan'],
+            ['code' => 'FM', 'name' => 'Micronesia'],
+            ['code' => 'MN', 'name' => 'Minnesota'],
+            ['code' => 'MS', 'name' => 'Mississippi'],
+            ['code' => 'MO', 'name' => 'Missouri'],
+            ['code' => 'MT', 'name' => 'Montana'],
+            ['code' => 'NE', 'name' => 'Nebraska'],
+            ['code' => 'NV', 'name' => 'Nevada'],
+            ['code' => 'NH', 'name' => 'New Hampshire'],
+            ['code' => 'NJ', 'name' => 'New Jersey'],
+            ['code' => 'NM', 'name' => 'New Mexico'],
+            ['code' => 'NY', 'name' => 'New York'],
+            ['code' => 'NC', 'name' => 'North Carolina'],
+            ['code' => 'ND', 'name' => 'North Dakota'],
+            ['code' => 'MP', 'name' => 'Northern Marianas'],
+            ['code' => 'OH', 'name' => 'Ohio'],
+            ['code' => 'OK', 'name' => 'Oklahoma'],
+            ['code' => 'OR', 'name' => 'Oregon'],
+            ['code' => 'PW', 'name' => 'Palau'],
+            ['code' => 'PA', 'name' => 'Pennsylvania'],
+            ['code' => 'PR', 'name' => 'Puerto Rico'],
+            ['code' => 'RI', 'name' => 'Rhode Island'],
+            ['code' => 'SC', 'name' => 'South Carolina'],
+            ['code' => 'SD', 'name' => 'South Dakota'],
+            ['code' => 'TN', 'name' => 'Tennessee'],
+            ['code' => 'TX', 'name' => 'Texas'],
+            ['code' => 'UT', 'name' => 'Utah'],
+            ['code' => 'VT', 'name' => 'Vermont'],
+            ['code' => 'VA', 'name' => 'Virginia'],
+            ['code' => 'VI', 'name' => 'Virgin Islands'],
+            ['code' => 'WA', 'name' => 'Washington'],
+            ['code' => 'WV', 'name' => 'West Virginia'],
+            ['code' => 'WI', 'name' => 'Wisconsin'],
+            ['code' => 'WY', 'name' => 'Wyoming']
+        ]);
+    }
+}
